proyecto_masking/requirements.txt
colorama==0.4.6
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.1
python-dotenv==1.2.1
tenacity==9.1.2
